refactor: reorganize model folder
