+++ data/jobpostings.json
@@ -19,7 +19,7 @@
     "contract_time": "full_time",
     "created": "2024-06-26T09:45:20Z",
     "redirect_url": "https://www.adzuna.com/details/4753284520?utm_medium=api&utm_source=2489f9d0",
-    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiNDc1MzI4NDUyMCIsInMiOiJFSWMzbWtWTTd4R2ZBNk0xMHZhcWR3In0.oRbifBjEU61PPrkfHtGPUuvi-1Z2eJDvVHCGLk09KR0",
+    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJzIjoiVmdybDMwZE03eEdsUU9VVzFyVXk2dyIsImkiOiI0NzUzMjg0NTIwIn0.W16DGn8Gi-zkEUbKwRqk507FgIMmaA4pGrFbOkzOJVA",
     "category": {
       "tag": "it-jobs",
       "label": "IT Jobs"
@@ -49,7 +49,7 @@
     "contract_time": "full_time",
     "created": "2022-07-30T22:02:27Z",
     "redirect_url": "https://www.adzuna.com/details/3361315315?utm_medium=api&utm_source=2489f9d0",
-    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiMzM2MTMxNTMxNSIsInMiOiJFSWMzbWtWTTd4R2ZBNk0xMHZhcWR3In0.VASoOucY7YXLlhue21gwoz8JuISYFGrIaQZRwR_m2E8",
+    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiMzM2MTMxNTMxNSIsInMiOiJWZ3JsMzBkTTd4R2xRT1VXMXJVeTZ3In0.oYX40sVAKsYzmWUcLfgLq7URMXiDPAgrgWvz1AxOoQQ",
     "category": {
       "tag": "it-jobs",
       "label": "IT Jobs"
@@ -79,7 +79,7 @@
     "contract_time": "full_time",
     "created": "2024-06-26T09:44:07Z",
     "redirect_url": "https://www.adzuna.com/details/4753282406?utm_medium=api&utm_source=2489f9d0",
-    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiNDc1MzI4MjQwNiIsInMiOiJFSWMzbWtWTTd4R2ZBNk0xMHZhcWR3In0.5A9moXbColZE-8iV1eDyHhN0A4KrgarKXPjXACDb14Q",
+    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiNDc1MzI4MjQwNiIsInMiOiJWZ3JsMzBkTTd4R2xRT1VXMXJVeTZ3In0.gSvZmP7VhwzzWYwmD9MILVEVW991-k4OBxf1IVKB0Ik",
     "category": {
       "tag": "it-jobs",
       "label": "IT Jobs"
@@ -108,8 +108,8 @@
     "salary_max": 119870.6,
     "contract_time": "full_time",
     "created": "2024-06-23T05:35:17Z",
-    "redirect_url": "https://www.adzuna.com/land/ad/4749359058?se=EIc3mkVM7xGfA6M10vaqdw&utm_medium=api&utm_source=2489f9d0&v=52984FFB98F14C3D779DE65F9C0C8035ACF7E4C5",
-    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJzIjoiRUljM21rVk03eEdmQTZNMTB2YXFkdyIsImkiOiI0NzQ5MzU5MDU4In0.U1AGzWPseDmUFHGxb6qgI2Ebwfkqb3opZjbQY9uh1oc",
+    "redirect_url": "https://www.adzuna.com/land/ad/4749359058?se=Vgrl30dM7xGlQOUW1rUy6w&utm_medium=api&utm_source=2489f9d0&v=52984FFB98F14C3D779DE65F9C0C8035ACF7E4C5",
+    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiNDc0OTM1OTA1OCIsInMiOiJWZ3JsMzBkTTd4R2xRT1VXMXJVeTZ3In0.LwtEYHuksX7A5rARv_xX8dk_W9fhR1C-qLHjhPXpH88",
     "category": {
       "tag": "it-jobs",
       "label": "IT Jobs"
@@ -139,7 +139,7 @@
     "contract_time": "full_time",
     "created": "2023-02-11T22:04:04Z",
     "redirect_url": "https://www.adzuna.com/details/3920756688?utm_medium=api&utm_source=2489f9d0",
-    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiMzkyMDc1NjY4OCIsInMiOiJFSWMzbWtWTTd4R2ZBNk0xMHZhcWR3In0.wrIn9hm0e7Kh0TjEVomBXmfP9FBNNb1wZKBxVbjOeA4",
+    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJzIjoiVmdybDMwZE03eEdsUU9VVzFyVXk2dyIsImkiOiIzOTIwNzU2Njg4In0.GDjOvVSxYNdWfX3L8d59qwNRKUXKWlbuBU-T2s3fBfo",
     "category": {
       "tag": "it-jobs",
       "label": "IT Jobs"
@@ -169,7 +169,7 @@
     "contract_time": "full_time",
     "created": "2024-04-14T03:48:19Z",
     "redirect_url": "https://www.adzuna.com/details/4648168464?utm_medium=api&utm_source=2489f9d0",
-    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiNDY0ODE2ODQ2NCIsInMiOiJFSWMzbWtWTTd4R2ZBNk0xMHZhcWR3In0.r9Wsn7-wwRYDmMHejZe88deEKT0IoBotJyiOFfZRtB8",
+    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiNDY0ODE2ODQ2NCIsInMiOiJWZ3JsMzBkTTd4R2xRT1VXMXJVeTZ3In0.WvkcS5OANiqUmcADX6vv6DPXX1xp6AHb_VJYfkGgzGk",
     "category": {
       "tag": "it-jobs",
       "label": "IT Jobs"
@@ -199,7 +199,7 @@
     "contract_time": "full_time",
     "created": "2024-07-25T01:17:48Z",
     "redirect_url": "https://www.adzuna.com/details/4794338227?utm_medium=api&utm_source=2489f9d0",
-    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiNDc5NDMzODIyNyIsInMiOiJFSWMzbWtWTTd4R2ZBNk0xMHZhcWR3In0.pFCKp2SJ43UbaExHY3jsGHX_NCDS6UNdE5XyBvytzr4",
+    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiNDc5NDMzODIyNyIsInMiOiJWZ3JsMzBkTTd4R2xRT1VXMXJVeTZ3In0.0J1dR4Q1-1q-JxcZwLySxvRlbxrSEyNSLCOBHRMU93k",
     "category": {
       "tag": "it-jobs",
       "label": "IT Jobs"
@@ -229,7 +229,7 @@
     "contract_time": "full_time",
     "created": "2022-02-24T20:11:46Z",
     "redirect_url": "https://www.adzuna.com/details/2924468724?utm_medium=api&utm_source=2489f9d0",
-    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJzIjoiRUljM21rVk03eEdmQTZNMTB2YXFkdyIsImkiOiIyOTI0NDY4NzI0In0.t9xkww3WZHdIvxvQEXnAvDenR-6-oClkoU6kSCQG7sI",
+    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiMjkyNDQ2ODcyNCIsInMiOiJWZ3JsMzBkTTd4R2xRT1VXMXJVeTZ3In0.laxO6zVcFemVBJbSn1SxTCsmKuGl769-l5cAbteBERo",
     "category": {
       "tag": "it-jobs",
       "label": "IT Jobs"
@@ -259,7 +259,7 @@
     "contract_time": "full_time",
     "created": "2024-06-01T23:35:22Z",
     "redirect_url": "https://www.adzuna.com/details/4717900686?utm_medium=api&utm_source=2489f9d0",
-    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiNDcxNzkwMDY4NiIsInMiOiJFSWMzbWtWTTd4R2ZBNk0xMHZhcWR3In0.dNsm5f2Q6RGiG5NdZeR8fNoOrJdcow0lIpA21V2rvbg",
+    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiNDcxNzkwMDY4NiIsInMiOiJWZ3JsMzBkTTd4R2xRT1VXMXJVeTZ3In0.SNrZW11wX8R527SQ_mJTvbtUBrp2jmdzxpZ99fYkHbc",
     "category": {
       "tag": "it-jobs",
       "label": "IT Jobs"
@@ -289,7 +289,7 @@
     "contract_time": "full_time",
     "created": "2024-07-04T00:57:15Z",
     "redirect_url": "https://www.adzuna.com/details/4765581889?utm_medium=api&utm_source=2489f9d0",
-    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiNDc2NTU4MTg4OSIsInMiOiJFSWMzbWtWTTd4R2ZBNk0xMHZhcWR3In0.jENdiy-F1Bd9PQiqemzrklH7WH5gCZOjQA5Tvmny33E",
+    "adref": "eyJhbGciOiJIUzI1NiJ9.eyJpIjoiNDc2NTU4MTg4OSIsInMiOiJWZ3JsMzBkTTd4R2xRT1VXMXJVeTZ3In0.WD5Sf4FgFsFEaqmlYeyTUBgngj9RyVi2NtpPMVAqSww",
     "category": {
       "tag": "it-jobs",
       "label": "IT Jobs"
