+++ src/main/java/jobplanner/model/api/JobPostUtil.java
@@ -1,10 +1,9 @@
-package jobplanner.model;
+package jobplanner.model.api;
 
 import java.io.InputStream;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.util.HashMap;
-import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -16,11 +15,9 @@ import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.xml.crypto.Data;
-
-import jobplanner.model.IJobPostModel.JobRecord;
 import jobplanner.model.formatters.DataFormatter;
 import jobplanner.model.formatters.Formats;
+import jobplanner.model.models.IJobPostModel.JobRecord;
 
 /**
  * A class to help with pulling data from
@@ -29,15 +26,22 @@ import jobplanner.model.formatters.Formats;
  * You can read more about the API at https://developer.adzuna.com/docs/search.
  */
 public final class JobPostUtil {
-    /** A Client Wrapper for Adzuna API */
+    /* API base url */
     private static String BASE_URL = "https://api.adzuna.com/v1/api/jobs/%s/%s/";
+    /* API id */
     private static String APP_ID;
+    /* API key */
     private static String APP_KEY;
+    /* Default country */
     private static String COUNTRY = "us";
+    /* Search endpoint */
     private static String SEARCH_ENDPOINT = "search";
 
     /**
-     * Prevent instantiation.
+     * Constructor for the JobPostUtil.
+     * 
+     * @param appId  The API id.
+     * @param appKey The API key.
      */
     public JobPostUtil(String appId, String appKey) {
         APP_ID = appId;
@@ -81,18 +85,42 @@ public final class JobPostUtil {
         return query.toString();
     }
 
+    /**
+     * Returns the job postings as an InputStream.
+     * 
+     * @return The job postings as an InputStream.
+     */
     public InputStream getJobPostings() {
         return getJobPostings(COUNTRY, null);
     }
 
+    /**
+     * Returns the job postings as an InputStream.
+     * 
+     * @param country The country to search in.
+     * @return The job postings as an InputStream.
+     */
     public InputStream getJobPostings(String country) {
         return getJobPostings(country, null);
     }
 
+    /**
+     * Returns the job postings as an InputStream.
+     * 
+     * @param params The parameters to search for.
+     * @return The job postings as an InputStream.
+     */
     public InputStream getJobPostings(HashMap<String, String> params) {
         return getJobPostings(COUNTRY, params);
     }
 
+    /**
+     * Returns the job postings as an InputStream.
+     * 
+     * @param country The country to search in.
+     * @param params  The parameters to search for.
+     * @return The job postings as an InputStream.
+     */
     public InputStream getJobPostings(String country, HashMap<String, String> params) {
         String query = buildQueryString(SEARCH_ENDPOINT, country, params);
         return getUrlContents(query);
@@ -106,7 +134,8 @@ public final class JobPostUtil {
      *         the connection
      *         fails
      * 
-     * Sourced from https://github.com/Su24-CS5004-Online-Lionelle/homework-07-Scarvy/blob/142c8726043fb851b0970b0955029db71d396662/src/main/java/student/model/net/NetUtils.java#L76
+     *         Sourced from
+     *         https://github.com/Su24-CS5004-Online-Lionelle/homework-07-Scarvy/blob/142c8726043fb851b0970b0955029db71d396662/src/main/java/student/model/net/NetUtils.java#L76
      */
     public InputStream getUrlContents(String urlStr) {
         try {
@@ -132,6 +161,11 @@ public final class JobPostUtil {
         return InputStream.nullInputStream();
     }
 
+    /**
+     * Main method to test the API request.
+     * 
+     * @param args The command line arguments.
+     */
     public static void main(String[] args) {
         // test the API request
         HashMap<String, String> params = new HashMap<>();
@@ -168,12 +202,8 @@ public final class JobPostUtil {
                 e.printStackTrace();
             }
         }
-        
-        // write the list of jobs to a file in JSON format using outputstream 
-        try {
-            DataFormatter.write(jobs, Formats.JSON, new FileOutputStream("data/jobpostings.json"));
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+
+        // write the list of jobs to a file in JSON format using outputstream
+        DataFormatter.write(jobs, Formats.JSON, System.out);
     }
 }
