+++ src/main/java/jobplanner/model/types/JobQueryParameter.java
@@ -1,6 +1,10 @@
-package jobplanner.model;
+package jobplanner.model.types;
 
+/**
+ * A list of job query parameters.
+ */
 public enum JobQueryParameter {
+    // The parameters for the job query
     WHAT("what"),
     WHERE("where"),
     COUNTRY("location0"),
@@ -14,12 +18,23 @@ public enum JobQueryParameter {
     PART_TIME("part_time"),
     CONTRACT("contract");
 
+    /** The tag for the parameter. */
     private final String tag;
 
+    /**
+     * Constructor for the job query parameter.
+     * 
+     * @param tag the tag for the parameter
+     */
     JobQueryParameter(String tag) {
         this.tag = tag;
     }
 
+    /**
+     * Gets the tag for the parameter.
+     * 
+     * @return the tag for the parameter
+     */
     public String getTag() {
         return tag;
     }
@@ -29,6 +44,12 @@ public enum JobQueryParameter {
         return tag;
     }
 
+    /**
+     * Gets the parameter from the tag.
+     * 
+     * @param tag the tag to get the parameter from
+     * @return the parameter
+     */
     public static JobQueryParameter fromTag(String tag) {
         for (JobQueryParameter parameter : JobQueryParameter.values()) {
             if (parameter.toString().equals(tag)) {
@@ -38,6 +59,12 @@ public enum JobQueryParameter {
         return null;
     }
 
+    /**
+     * Gets the parameter from the string.
+     * 
+     * @param value the string to get the parameter from
+     * @return the parameter
+     */
     public static JobQueryParameter fromString(String value) {
         for (JobQueryParameter parameter : JobQueryParameter.values()) {
             if (parameter.toString().equals(value)) {
